Add support for single children
Up until now, I've been working on the assumption that FlipMove will
wrap around an array of elements. This assumption was challenged while
building out the demos, as I wanted a single element to move around
depending on state.

The problem is that React makes no attempt to convert single children
to an array, so I had to do that manually, using the
React.Children.toArray helper.

This release also includes a bunch of refactoring and improvements.