Filter out unchanged children.
Previously, the initial filter was just to catch items that had been
created/destroyed in this state, or items that didn't have a key.

The problem is, that means all items whose positions haven't changed
are passed along to the animate function. And while those with no
delta were returned early, their index still counted.

If FlipMove with 9 elements has a `staggerDurationBy` of 40ms, that
means there are 9 stages of staggered duration, _no matter how many
of the children have moved_. if only the last child has budged, it
means it would take 320ms longer than it was supposed to.